+++ src/tests/integration/components/cluster-host-ip-form/host-row-test.js
@@ -8,11 +8,11 @@ describe('Integration | Component | cluster-host-ip-form/host-row', function () 
   setupRenderingTest();
 
   it('renders static text if readonly flag is true', async function () {
-    await render(hbs `{{cluster-host-ip-form/host-row
-      isReadOnly=true
-      hostname="hostname.one"
-      ip="172.17.0.1"
-    }}`);
+    await render(hbs `<ClusterHostIpForm::HostRow
+      @isReadOnly={{true}}
+      @hostname="hostname.one"
+      @ip="172.17.0.1"
+    />`);
 
     expect(find('input')).to.not.exist;
     expect(find('.cluster-host-ip-form-row'))
@@ -20,11 +20,11 @@ describe('Integration | Component | cluster-host-ip-form/host-row', function () 
   });
 
   it('renders input if readonly flag is false', async function () {
-    await render(hbs `{{cluster-host-ip-form/host-row
-      isReadOnly=false
-      hostname="hostname.one"
-      ip="172.17.0.1"
-    }}`);
+    await render(hbs `<ClusterHostIpForm::HostRow
+      @isReadOnly={{false}}
+      @hostname="hostname.one"
+      @ip="172.17.0.1"
+    />`);
 
     const input = find('input');